requirements.txt
langchain
langchain-openai
langchain_community
langchain_core
python-dotenv
streamlit
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
langchain_chroma
duckdb
pandas
openai
langchain_groq